Refactor serializers into separate files
