+++ src/meshapi/serializers/map.py
@@ -1,34 +1,9 @@
 import datetime
 from collections import OrderedDict
 
-from django.contrib.auth.models import User
 from rest_framework import serializers
 
-from meshapi.models import Building, Install, Link, Member, Sector
-
-
-class UserSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = User
-        fields = "__all__"
-
-
-class BuildingSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = Building
-        fields = "__all__"
-
-
-class MemberSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = Member
-        fields = "__all__"
-
-
-class InstallSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = Install
-        fields = "__all__"
+from meshapi.models import Install, Link, Sector
 
 
 class JavascriptDateField(serializers.IntegerField):
